% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysore2015.R
\docType{data}
\name{mysore2015exp}
\alias{mysore2015exp}
\title{Explanatory variables for the study}
\format{A data frame with 102 rows and 28 variables:
\describe{
  \item{ID}{Subject ID}
  \item{nausea}{Current nausea (0 = no, 1 = yes).}
  \item{vomit}{Current vomiting (0 = no, 1 = yes).}
  \item{HWWStoilet}{do you hand wash with soap after using toilet? (0 = no, 1 = yes).}
  \item{HWWScleanchild}{do you hand wash with soap after cleaning child? (0 = no, 1 = yes).}
  \item{HWWSfeedchild}{do you hand wash with soap before feeding child? (0 = no, 1 = yes).}
  \item{HWWSfood}{do you hand wash with soap before handling food? (0 = no, 1 = yes).}
  \item{immune1}{In general, I am very susceptible to colds, flu and other infectious diseases (0 = no, 1 = yes).}
  \item{immune2}{My immune system protects me from most illnesses that other people get (0 = no, 1 = yes).}
  \item{immune3}{I have a history of susceptibility to infectious disease (0 = no, 1 = yes).}
  \item{toilet}{type of toilet: none (0) or squat (1)}
  \item{waterclean}{perceived cleanliness of drinking water: poor (0), good (1).}
  \item{dietadvice}{who do you ask for advice regarding diet in pregnancy}
  \item{dietadvice2}{who do you ask for advice regarding diet in pregnancy}
  \item{dietadvice3}{who do you ask for advice regarding diet in pregnancy}
  \item{dietadvice4}{who do you ask for advice regarding diet in pregnancy}
  \item{pressure}{does anyone pressure to follow certain guidelines in pregnancy? (first person listed)}
  \item{pressure2}{2nd person listed that gives pressure to follow pregnancy guidelines}
  \item{pressure3}{3rd person listed that gives pressure to follow pregnancy guiedlines}
  \item{totalhousehold}{total number of individuals living in house}
  \item{friend.sis.pregnantYN}{Are any of your friends or sisters pregnant?}
  \item{fs1}{food insecurity item 1}
  \item{fs2}{food insecurity item 2}
  \item{fs3}{food insecurity item 3}
  \item{fs4}{food insecurity item 4}
  \item{fs5}{food insecurity item 5}
  \item{fs6}{food insecurity item 6}
  \item{fstotal}{total food insecurity score}
}}
\source{
\url{placekmysore2015publicdata}
}
\usage{
mysore2015exp
}
\description{
A dataset containing the explanatory variables, such as food insecurity, perceived susceptibility to
disease, and so forth. Food insecurity items are from the 6-item short-form household-level food insecurity measure, 
which assesses one’s access to sufficient foods (Blumberg, Bialostosky, Hamilton, & Briefel, 1999).
}
\keyword{datasets}
