% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysore2015.R
\docType{data}
\name{mysore2015food}
\alias{mysore2015food}
\title{Free-listed food avoidances and aversions}
\format{A data frame with 102 rows and 69 variables:
\describe{
  \item{ID}{participant ID}
  \item{aversion}{first item listed as pregnancy aversion}
  \item{symptom}{symptom experienced from first mentioned aversion}
  \item{aversion2}{second item listed as pregnancy aversion}
  \item{symptom2}{symptom experienced from second mentioned aversion}
  \item{aversion3}{third item listed as pregnancy aversion}
  \item{symptom3}{symptom experienced from third mentioned aversion}
  \item{aversion4}{fourth item listed as pregnancy aversion}
  \item{symptom4}{symptom experienced from fourth mentioned aversion}
  \item{aversion5}{fifth item listed as pregnancy aversion}
  \item{symptom5}{symptom experienced from fifth mentioned aversion}
  \item{avoidance1}{first item listed as pregnancy avoidance}
  \item{learn1}{how individual learned about food avoidance}
  \item{consequence1}{consequence of consuming forbidden item}
  \item{avoidance2}{second item listed as pregnancy avoidance}
  \item{learn2}{how individual learned about food avoidance}
  \item{consequence2}{consequence of consuming forbidden item}
  \item{avoidance3}{third item listed as pregnancy avoidance}
  \item{learn3}{how individual learned about food avoidance}
  \item{consequence3}{consequence of consuming forbidden item}
  \item{avoidance4}{fourth item listed as pregnancy avoidance}
  \item{learn4}{how individual learned about food avoidance}
  \item{consequence4}{consequence of consuming forbidden item}
  \item{avoidance5}{fifth item listed as pregnancy avoidance}
  \item{learn5}{how individual learned about food avoidance}
  \item{consequence5}{consequence of consuming forbidden item}
  \item{avoidance6}{sixth item listed as pregnancy avoidance}
  \item{learn6}{how individual learned about food avoidance}
  \item{consequence6}{consequence of consuming forbidden item}
  \item{avoidance7}{seventh item listed as pregnancy avoidance}
  \item{learn7}{how individual learned about food avoidance}
  \item{consequence7}{consequence of consuming forbidden item}
  \item{\code{fruit_avoid}}{logical. Does the participant avoid at least one fruit?}
  \item{\code{meat_avoid}}{logical. Does the participant avoid at least one meat?}
  \item{\code{vegetable_avoid}}{logical. Does the participant avoid at least one vegetable?}
  \item{\code{nut.seed.legume_avoid}}{logical. Does the participant avoid at least one nut/seed/legume?}
  \item{\code{grains_avoid}}{logical. Does the participant avoid at least one grain?}
  \item{\code{ess_avoid}}{logical. Does the participant avoid at least one ethnic strong and spicy food?}
  \item{\code{dairy_avoid}}{logical. Does the participant avoid at least one dairy food?}
  \item{\code{fruit_averse}}{logical. Is the participant averse to at least one fruit?}
  \item{\code{meat_averse}}{logical. Is the participant averse to at least one meat?}
  \item{\code{vegetable_averse}}{logical. Is the participant averse to at least one vegetable?}
  \item{\code{nut.seed.legume_averse}}{logical. Is the participant averse to at least one nut/seed/legume?}
  \item{\code{ess_averse}}{logical. Is the participant averse to at least one ethnic strong and spicy food?}
  \item{\code{grains_averse}}{logical. Is the participant averse to at least one grain?}
  \item{\code{sweets_averse}}{logical. Is the participant averse to at least one sweet?}
}}
\source{
\url{https://github.com/grasshoppermouse/placekmysore2015publicdata}
}
\usage{
mysore2015food
}
\description{
A dataset containing free-listed food avoidances and aversions, the symptoms
experienced if consuming aversive foods, and the perceived consequences of
consuming foods that are taboo.
}
\keyword{datasets}
