% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysore2015.R
\docType{data}
\name{mysore2015photos}
\alias{mysore2015photos}
\title{Food photo ratings}
\format{A data frame with 102 rows and 63 variables:
\describe{
  \item{ID}{participant ID number}
  \item{mango1like}{current preference for mangoes}
  \item{mango.bplike1}{preference for mango before pregnancy}
  \item{sweets2like}{current preference for sweets}
  \item{sweets.bplike2}{preference for sweets before pregnancy}
  \item{coffee3like}{current preference for coffee}
  \item{coffee.bplike3}{preference for coffee before pregnancy}
  \item{unripemango4like}{current preference for unripe mango}
  \item{unriplemango.bplike4}{preference for unripe mango before pregnancy}
  \item{papaya.5like}{current preference for papaya}
  \item{papaya.bplike5}{preference for papaya before pregnancy}
  \item{blackgrapes6like}{current preference for black grapes}
  \item{blackgrapes.bplike6}{preference for black grapes before pregnancy}
  \item{bananas7like}{current preference for bananas}
  \item{bananas.bplike7}{preference for bananas before pregnancy}
  \item{jackfruit8like}{current preference for jackfruit}
  \item{jackfruit.bplike8}{preference for jackfruit before pregnancy}
  \item{mangopickle9like}{current preference for mango pickle}
  \item{mangopickle.bplike9}{preference for mango pickle before pregnancy}
  \item{rice10like}{current preference for rice}
  \item{rice.bplike10}{preference for rice before pregnancy}
  \item{sambar11like}{current preference for sambar}
  \item{sambar.bplike11}{preference for sambar before pregnancy}
  \item{sesame12like}{current preference for sesame}
  \item{sesame.bplike12}{preference for sesame before pregnancy}
  \item{salt13like}{current preference for salt}
  \item{salt.bplike13}{preference for salt before pregnancy}
  \item{coconut14like}{current preference for coconut}
  \item{coconut.bplike14}{preference for coconut before pregnancy}
  \item{pomegranite15like}{current preference for pomegranite}
  \item{pomegranite.bplike15}{preference for pomegranite before pregnancy}
  \item{uncookedrice16like}{current preference for uncooked rice}
  \item{uncookedrice.bplike16}{preference for uncooked rice before pregnancy}
  \item{tamarind17like}{current preference for tamarind}
  \item{tamarind.bplike17}{preference for tamarind before pregnancy}
  \item{guava19like}{current preference for guava}
  \item{guava.bplike19}{preference for guava before pregnancy}
  \item{horsegram20like}{current preference for horse gram}
  \item{horsegram.bplike20}{preference for horse gram before pregnancy}
  \item{eggplant21like}{current preference for eggplant}
  \item{eggplant.bplike21}{preference for eggplant before pregnancy}
  \item{milk22like}{current preference for milk}
  \item{milk.bplike22}{preference for milk before pregnancy}
  \item{drumstick23like}{current preference for drumstick}
  \item{drumstick.bplike23}{preference for drumstick before pregnancy}
  \item{curdrice24like}{current preference for curd rice}
  \item{curdrice.bplike24}{preference for curd rice before pregnancy}
  \item{egg25like}{current preference for egg}
  \item{egg.bplike25}{preference for egg before pregnancy}
  \item{dill26like}{current preference for dill}
  \item{dill.bplike26}{preference for dill before pregnancy}
  \item{fish27like}{current preference for fish}
  \item{fish.bplike27}{preference for fish before pregnancy}
  \item{bamboo28like}{current preference for bamboo}
  \item{bamboo.bplike28}{preference for bamboo before pregnancy}
  \item{spinach29like}{current preference for spinach}
  \item{spinach.bplike29}{preference for spinach before pregnancy}
  \item{potato30like}{current preference for potato}
  \item{potato.bplike30}{preference for potato before pregnancy}
  \item{chicken31like}{current preference for chicken}
  \item{chicken.bplike31}{preference for chicken before pregnancy}
  \item{bittergourd32like}{current preference for bitter gourd}
  \item{bittergourd.bplike32}{preference for bitter gourd before pregnancy}
}}
\source{
\url{https://github.com/grasshoppermouse/placekmysore2015publicdata}
}
\usage{
mysore2015photos
}
\description{
Ratings of 31 food photos before and during pregnancy. All data collected during pregnancy.
0 = dislike, 1 = sometimes like, 2 = like.
}
\keyword{datasets}
