% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placekmysore2015data.R
\docType{package}
\name{placekmysore2015data}
\alias{placekmysore2015data}
\alias{placekmysore2015data-package}
\title{Data on pregnancy diet in Mysore, India.}
\description{
Data on pregnancy diet in Mysore, India.
}
\details{
During pregnancy, many women report feelings of disgust toward ordinary foods,
and cravings for particular foods. The purpose of this study is to answer
questions about the socioecological, environmental, psychological, and
adaptive influences of such dietary shifts in pregnancy. Specifically,
this study aims to answer the following questions: 1) What role does social
learning (culture) and social norms have in the types of foods/non-foods that
are craved and avoided during pregnancy? 2) Do environmental factors, such as
resource scarcity (i.e. food insecurity) and pathogen risk, influence putative
dietary shifts in pregnancy? 3) What role does nausea and
vomiting in pregnancy (NVP) have in predicting dietary shifts?

Data were collected Summer 2015.
}
\seealso{
\code{\link{mysore2015dem}}, \code{\link{mysore2015exp}}, \code{\link{mysore2015food}}, \code{\link{mysore2015photos}}
}
\author{
\email{caitlyn.placek@gmail.com}
}
\keyword{package}
