% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysore2015.R
\docType{data}
\name{mysore2015exp}
\alias{mysore2015exp}
\title{Explanatory variables for the study}
\format{A data frame with 102 rows and 28 variables:
\describe{
  \item{ID}{Subject ID}
  \item{nausea}{Current nausea (0 = no, 1 = yes).}
  \item{vomit}{Current vomiting (0 = no, 1 = yes).}
  \item{HWWStoilet}{do you hand wash with soap after using toilet? (0 = no, 1 = yes).}
  \item{HWWScleanchild}{do you hand wash with soap after cleaning child? (0 = no, 1 = yes).}
  \item{HWWSfeedchild}{do you hand wash with soap before feeding child? (0 = no, 1 = yes).}
  \item{HWWSfood}{do you hand wash with soap before handling food? (0 = no, 1 = yes).}
  \item{immune1}{In general, I am very susceptible to colds, flu and other infectious diseases (0 = no, 1 = yes).}
  \item{immune2}{My immune system protects me from most illnesses that other people get (0 = no, 1 = yes).}
  \item{immune3}{I have a history of susceptibility to infectious disease (0 = no, 1 = yes).}
  \item{toilet}{type of toilet: none (0) or squat (1)}
  \item{waterclean}{perceived cleanliness of drinking water: poor (0), good (1).}
  \item{dietadvice}{who do you ask for advice regarding diet in pregnancy}
  \item{dietadvice2}{who do you ask for advice regarding diet in pregnancy}
  \item{dietadvice3}{who do you ask for advice regarding diet in pregnancy}
  \item{dietadvice4}{who do you ask for advice regarding diet in pregnancy}
  \item{pressure}{does anyone pressure to follow certain guidelines in pregnancy? (first person listed)}
  \item{pressure2}{2nd person listed that gives pressure to follow pregnancy guidelines}
  \item{pressure3}{3rd person listed that gives pressure to follow pregnancy guiedlines}
  \item{totalhousehold}{total number of individuals living in house.  This variable is omitted to maintain anonymity, but will be provided to researchers who agree to the terms of our IRB.}
  \item{friend.sis.pregnantYN}{Are any of your friends or sisters pregnant?}
  \item{fs1}{In the last 12 months, since (date 12 months ago), did you (or other adults in your household) ever cut the size of your meals or skip meals because there wasn't enough money for food? (0 = no, 1 = yes)}
  \item{fs2}{[Ask only if answered YES to fs1] How often did this happen? (0 = never; 1 = Some months but not every month; 2 = Almost every month)}
  \item{fs3}{In the last 12 months, did you ever eat less than you felt you should because there wasn't enough money to buy food? (0 = no, 1 = yes)}
  \item{fs4}{In the last 12 months, since (date 12 months ago), were you ever hungry but didn't eat because you couldn't afford enough food? (0 = no, 1 = yes)}
  \item{fs5}{'The food that [I/we] bought just didn't last, and [I/we] didn't have money to get more.' Was that often, sometimes, or never true for you in the last 12 months? (0 = never, 1 = sometimes, 2 = often)}
  \item{fs6}{'[I/we] couldn't afford to eat balanced meals.' Was that often, sometimes, or never true for you in the last 12 months? (0 = never, 1 = sometimes, 2 = often)}
  \item{fstotal}{total food insecurity score computed per Blumberg, Bialostosky, Hamilton, & Briefel (1999)}
}}
\source{
\url{placekmysore2015publicdata}
}
\usage{
mysore2015exp
}
\description{
A dataset containing the explanatory variables, such as food insecurity, perceived susceptibility to
disease, and so forth. Food insecurity items are from the 6-item short-form household-level food insecurity measure, 
which assesses one’s access to sufficient foods (Blumberg, Bialostosky, Hamilton, & Briefel, 1999).
}
\keyword{datasets}
